refactor(core,control-interface): InterfaceLinkDefinition -> core
This commit refators the types defined in both `wasmcloud-core` and
`wasmcloud-control-interface` to make it easier to distinguish what
types belong where and what they're related to.

Ultimately the goal here was was to move `InterfaceLinkDefinition`
into `wasmcloud-core` so it can be used in other places, but it was a
good chance to reorganize.
