+++ crates/control-interface/src/types/mod.rs
@@ -1,0 +1,9 @@
+//! Collection of types that are commonly used/necessary in control interface operations
+
+pub mod actor;
+pub mod ctl;
+pub mod host;
+pub mod link;
+pub mod provider;
+pub mod registry;
+pub mod rpc;
