+++ crates/core/src/rpc.rs
@@ -1,0 +1,14 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct HealthCheckRequest {}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct HealthCheckResponse {
+    /// A flag that indicates the the actor is healthy
+    #[serde(default)]
+    pub healthy: bool,
+    /// A message containing additional information about the actors health
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+}
